Share credential generation logic across multiple modules
Refactors credential generation functions into a shared resource module and updates dependent modules to import and utilize these functions, removing redundant code.
