+++ cmd/webhook/handlers/tenant-composite-sync.js
@@ -31,7 +31,13 @@
  */
 
 const crypto = require('crypto');
-const { ANNOTATIONS, LABELS } = require('./constants');
+const { ANNOTATIONS, LABELS, RESOURCE_TYPES } = require('./constants');
+const { 
+  generateSecurePassword, 
+  generateUsername,
+  generateAppCredentials,
+  renderAppCredentialsSecret 
+} = require('./resources');
 
 function createTenantCompositeSync(config) {
   
@@ -788,83 +794,6 @@ function createTenantCompositeSync(config) {
     renderChildren,
     sendResponse
   ];
-}
-
-
-function renderAppCredentialsSecret(tenantId, namespace, databaseName, existingSecret, labels) {
-      
-  const secretName = `${tenantId}-app-credentials`;
-  var existingAnnotations = existingSecret?.metadata?.annotations || { };
-  var existingLabels = existingSecret?.metadata?.labels || { };
-
-  // Standard annotations that are always set
-  const standardAnnotations = {
-    'ns.mdn.io/created-at': new Date().toISOString(),
-    'ns.mdn.io/tenant': tenantId,
-    'ns.mdn.io/description': 'MongoDB credentials for Nightscout application pods',
-    [ANNOTATIONS.PROTECTED_RESOURCE]: 'true'
-  };
-  
-  // Merge existing annotations (lifecycle tracking) with standard annotations
-  // Existing annotations take precedence to preserve user-initialized state
-  const mergedAnnotations = {
-    ...standardAnnotations,
-    ...existingAnnotations
-  };
-
-  const mergedLabels = {
-    ...labels,
-    ...existingLabels
-  };
-
-  var secret = {
-    apiVersion: 'v1',
-    kind: 'Secret',
-    metadata: {
-      name: secretName,
-      namespace: namespace,
-      labels: {
-        ...mergedLabels,
-        'app.kubernetes.io/component': 'app-credentials',
-        'app.kubernetes.io/managed-by': 'metacontroller',
-        'ns.mdn.io/decorator': 'storage-credentials',
-        'ns.mdn.io/credential-type': 'application',
-        [LABELS.RESOURCE_TYPE]: RESOURCE_TYPES.APP_CREDENTIALS_SECRET
-      },
-      annotations: mergedAnnotations
-    },
-    type: 'Opaque',
-
-  };
-
-  if (existingSecret) {
-    secret.data = existingSecret.data;
-    secret.type = existingSecret.type;
-  } else {
-    console.log(`  FIRST CYCLE - GENERATING NEW APP CREDENTIALS`);
-    const username = generateUsername(tenantId);
-    const password = generateSecurePassword(16);
-    
-    const mongoHost = `mongo-${databaseName}`;
-    const mongoPort = '27017';
-    
-    const appCredentials = generateAppCredentials(
-      tenantId,
-      mongoHost,
-      mongoPort,
-      databaseName,
-      username,
-      password
-    );
-
-    // Encode all credential fields to base64
-    var encodedData = {};
-    Object.keys(appCredentials).forEach(key => {
-      encodedData[key] = Buffer.from(appCredentials[key]).toString('base64');
-    });
-    secret.data = encodedData;
-  }
-  return secret;
 }
 
 module.exports = createTenantCompositeSync;
